requirements.txt
python-dotenv
openai
selenium
webdriver-manager
diskcache
pytube
